Move funding states
