+++ packages/wallet/src/redux/channel/funding/sharedState.ts
